model_ai/requirements.txt
flask[async]
ultralytics
pillow
gunicorn
uvicorn